chore(vwo): move sdk loader to separate file
